[React Redux] Move selector out from container
